+++ packages/components-js/tests/vrt/playwright/specs/states/modal.vrt.ts
@@ -3,6 +3,7 @@ import { schemes, themes, viewportWidthM } from '@porsche-design-system/shared/t
 import {
   forceFocusHoverState,
   forceFocusState,
+  forceFocusVisibleState,
   forceHoverState,
   getPlaygroundPseudoStatesMarkup,
   type PrefersColorScheme,
@@ -38,9 +39,9 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
   // due to custom hover state we need to set hover also on component itself
   await forceHoverState(page, '.hover p-modal >>> p-button-pure');
   await forceHoverState(page, '.hover p-modal >>> p-button-pure >>> button');
-  await forceFocusState(page, '.focus p-modal a');
+  await forceFocusVisibleState(page, '.focus p-modal a');
   await forceFocusState(page, '.focus p-modal >>> .root');
-  await forceFocusState(page, '.focus p-modal >>> p-button-pure >>> button');
+  await forceFocusVisibleState(page, '.focus p-modal >>> p-button-pure >>> button');
   await forceFocusHoverState(page, '.focus-hover p-modal a');
   await forceFocusHoverState(page, '.focus-hover p-modal >>> .root');
   // due to custom hover state we need to set hover also on component itself