+++ packages/components-js/tests/vrt/playwright/specs/states/select.vrt.ts
@@ -3,6 +3,7 @@ import { schemes, themes, viewportWidthM } from '@porsche-design-system/shared/t
 import {
   forceFocusHoverState,
   forceFocusState,
+  forceFocusVisibleState,
   forceHoverState,
   getPlaygroundPseudoStatesMarkup,
   type PrefersColorScheme,
@@ -124,7 +125,7 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
 
   await forceHoverState(page, '.hover p-select >>> button');
   await forceHoverState(page, '.hover p-select span a');
-  await forceFocusState(page, '.focus p-select span a');
+  await forceFocusVisibleState(page, '.focus p-select span a');
   await forceFocusState(page, '.focus p-select >>> button');
   await forceFocusHoverState(page, '.focus-hover p-select >>> button');
   await forceFocusHoverState(page, '.focus-hover p-select span a');