+++ packages/components-js/tests/vrt/playwright/specs/states/switch.vrt.ts
@@ -2,7 +2,7 @@ import { expect, type Page, test } from '@playwright/test';
 import { schemes, themes, viewportWidthM } from '@porsche-design-system/shared/testing/playwright.vrt.config';
 import {
   forceFocusHoverState,
-  forceFocusState,
+  forceFocusVisibleState,
   forceHoverState,
   getPlaygroundPseudoStatesMarkup,
   type PrefersColorScheme,
@@ -32,9 +32,9 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
 
   await forceHoverState(page, '.hover p-switch >>> button');
   await forceHoverState(page, '.hover p-switch span a');
-  await forceFocusState(page, '.focus p-switch'); // native outline should not be visible
-  await forceFocusState(page, '.focus p-switch >>> button');
-  await forceFocusState(page, '.focus p-switch span a');
+  await forceFocusVisibleState(page, '.focus p-switch'); // native outline should not be visible
+  await forceFocusVisibleState(page, '.focus p-switch >>> button');
+  await forceFocusVisibleState(page, '.focus p-switch span a');
   await forceFocusHoverState(page, '.focus-hover p-switch >>> button');
   await forceFocusHoverState(page, '.focus-hover p-switch span a');
 };