+++ packages/components-js/tests/vrt/playwright/specs/states/text.vrt.ts
@@ -2,7 +2,7 @@ import { expect, type Page, test } from '@playwright/test';
 import { schemes, themes, viewportWidthM } from '@porsche-design-system/shared/testing/playwright.vrt.config';
 import {
   forceFocusHoverState,
-  forceFocusState,
+  forceFocusVisibleState,
   forceHoverState,
   getPlaygroundPseudoStatesMarkup,
   type PrefersColorScheme,
@@ -28,8 +28,8 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
 
   await forceHoverState(page, '.hover p-text a');
   await forceHoverState(page, '.hover p-text button');
-  await forceFocusState(page, '.focus p-text a');
-  await forceFocusState(page, '.focus p-text button');
+  await forceFocusVisibleState(page, '.focus p-text a');
+  await forceFocusVisibleState(page, '.focus p-text button');
   await forceFocusHoverState(page, '.focus-hover p-text a');
   await forceFocusHoverState(page, '.focus-hover p-text button');
 };