+++ packages/components-vue/src/examples/LinkTileProductExample.vue
@@ -14,7 +14,7 @@
     :heading="'Some product name'"
     :price="'1.911,00 €'"
     :description="'Some description'"
-    :href="'https://www.porsche.com'"
+    :href="'https://porsche.com'"
     :liked="liked"
     @like="handleLike"
   >