+++ packages/components/src/utils/breakpoint-customizable.spec.ts
@@ -15,7 +15,7 @@ describe('parseJSON()', () => {
     ["{base: 'initial', l: 'equal'}", { base: 'initial', l: 'equal' }],
     ['{base: "initial", l: "equal"}', { base: 'initial', l: 'equal' }],
     ['{"base": "right", "s": "left"}', { base: 'right', s: 'left' }],
-    ['{base: "https://www.porsche.com"}', { base: 'https://www.porsche.com' }],
+    ['{base: "https://porsche.com"}', { base: 'https://porsche.com' }],
     ['{base: 36, xs: 6, s: 4, m: 16, l: 2}', { base: 36, xs: 6, s: 4, m: 16, l: 2 }],
     [
       '{base: "1:1", s: "3:4", m: "4:3", l: "9:16", xl: "16:9", xxl: "3:4"}',