+++ packages/storefront/src/components/Intro.vue
@@ -8,10 +8,10 @@
     </p-text>
     <p-text :theme="storefrontTheme" class="paragraph" color="contrast-high">
       We provide protected
-      <a href="https://www.figma.com/file/EkdP468u4ZVuIRwalKCscb/Design-System-v3?type=design&node-id=105-146">Figma</a>
+      <a href="https://figma.com/file/EkdP468u4ZVuIRwalKCscb/Design-System-v3?type=design&node-id=105-146">Figma</a>
       and UX Pin libraries, as well as public accessible
-      <a href="https://www.npmjs.com/org/porsche-design-system">npm releases</a> for Vanilla JS,
-      <a href="https://angular.io" target="_blank">Angular</a>, <a href="https://react.dev/">React</a>,
+      <a href="https://npmjs.com/org/porsche-design-system">npm releases</a> for Vanilla JS,
+      <a href="https://angular.io" target="_blank">Angular</a>, <a href="https://react.dev">React</a>,
       <a href="https://nextjs.org">Next.js</a>, <a href="https://remix.run">Remix</a> and
       <a href="https://vuejs.org">Vue</a> based on Web Components with TypeScript and SSR support and appreciate any
       feedback, feature requests or suggestions on our public accessible