+++ packages/storefront/tests/unit:vitest/config/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    include: ['**/tests/unit:vitest/specs/**/*.spec.ts'],
+    testTimeout: 10000,
+  },
+});